Prevent Enter keydown events from being fired in the modal.
Prior to WP 6.7.1 update, we had custom logic to ensure that Enter
keydowns inside the modal would not fire the default behavior,
except when focus was on the Add Attribution button. Something changed
in WordPress such that the event.preventDefault() call was too late.
WP was reclaiming focus, and Enter was performing the default action
in the editor rather than the modal. This resulted in the modal being
closed with no attribution, and a line break added to the editor.

Rather than debugging the entire event chain, we've added code
that disables Enter altogether in the modal. Users who are using
the keyboard to navigate will have to use the space bar to submit.

See http://redmine.citytech.cuny.edu/issues/3476.